requirements.txt
Flask
gunicorn
psycopg2-binary
python-dotenv
requests
Flask-SocketIO
python-telegram-bot
eventlet